feat(hil): nfsboot unzip rts
